backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.13.1

# Redis
redis>=5.0.1

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Pydantic and settings
pydantic>=2.10.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP client
httpx==0.26.0

# Date and time utilities
python-dateutil==2.8.2

# Scheduling (for future background tasks)
apscheduler==3.10.4

# Production server
gunicorn==21.2.0
